Migrate Rust GPIOd code into C code
The version of GPIOd used in Rust seems to no longer work due to kernel
side changes. Because this change has been also reflected in the
libgpiod library (thats written itself in C), we can just write C code
that iterfaces with the library instead and that way we remain the same
functionality in spite of the Rust lib being outdated.